requirements.txt
requests
django
djangorestframework
djangorestframework-simplejwt